+++ src/monmove.c
@@ -208,7 +208,7 @@ onscary(int x, int y, struct monst* mtmp)
 void
 mon_regen(struct monst* mon, boolean digest_meal)
 {
-    if (mon->mhp < mon->mhpmax && (g.moves % 20 == mon->m_id % 20
+    if (mon->mhp < mon->mhpmax && (g.moves % 20 == (long) mon->m_id % 20
                                    || regenerates(mon->data)))
         mon->mhp++;
     if (mon->mspec_used)
